Fix #34: Introduce fuzzyLocales option to redirect
This change introduces the fuzzyLocales option to the redirect plugin.
It allows users to configure the language-based redirects in a way that
they only take the language into account while ignoring specific
dialects. For example, this means that users with browsers that send
`de-AT` as a locale will still be forwarded to the `de-DE` locale, as
the plugin will only take the lowercase `de` into account while ignoring
the region-specific uppercase `DE` and `AT`. This is only a fallback, so
if both exists, users will be forwarded to the appropriate dialect
version of their language.