Fix #34: Redirect to lang if dialect is not found
This change means that users with browsers that send an unknow region of
a known language will still be forwarded to the correct overall language.
For example, users with a browser that sends `de-AT` as a locale
will still be forwarded to the `de-DE` locale, as the plugin will
only take the lowercase `de` into account while ignoring
the region-specific uppercase `DE` and `AT`.