fix(dehydrate): find string in a right way
`String.prototype.find` will treat the first argument as regexp.

The vairable `resourceOriginal`, which contents is `{{{ renderResourceHints() }}}`, was not works as expected because `()` was treated as regex match groups. This leads to many unused tags generated in the ssr template.

indexOf is safe for any string to find the neddle in string.